+++ lib/new_relic/agent/configuration/default_source.rb
@@ -1505,6 +1505,20 @@ module NewRelic
           :allowed_from_server => false,
           :description => 'If <code>true</code>, the agent captures attributes from browser monitoring.'
         },
+        :'span_events.attributes.enabled' => {
+          :default     => true,
+          :public      => true,
+          :type        => Boolean,
+          :allowed_from_server => false,
+          :description => 'If <code>true</code>, the agent captures attributes on span events.'
+        },
+        :'transaction_segments.attributes.enabled' => {
+          :default     => true,
+          :public      => true,
+          :type        => Boolean,
+          :allowed_from_server => false,
+          :description => 'If <code>true</code>, the agent captures attributes on transaction segments.'
+        },
         :'attributes.exclude' => {
           :default     => [],
           :public      => true,
@@ -1545,6 +1559,22 @@ module NewRelic
           :transform    => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to exclude from browser monitoring. Allows <code>*</code> as wildcard at end.'
         },
+        :'span_events.attributes.exclude' => {
+          :default     => [],
+          :public      => true,
+          :type        => Array,
+          :allowed_from_server => false,
+          :transform   => DefaultSource.method(:convert_to_list),
+          :description => 'Prefix of attributes to exclude from span events. Allows <code>*</code> as wildcard at end.'
+        },
+        :'transaction_segments.attributes.exclude' => {
+          :default     => [],
+          :public      => true,
+          :type        => Array,
+          :allowed_from_server => false,
+          :transform   => DefaultSource.method(:convert_to_list),
+          :description => 'Prefix of attributes to exclude from transaction segments. Allows <code>*</code> as wildcard at end.'
+        },
         :'attributes.include' => {
           :default     => [],
           :public      => true,
@@ -1584,6 +1614,22 @@ module NewRelic
           :allowed_from_server => false,
           :transform    => DefaultSource.method(:convert_to_list),
           :description => 'Prefix of attributes to include in browser monitoring. Allows <code>*</code> as wildcard at end.'
+        },
+        :'span_events.attributes.include' => {
+          :default     => [],
+          :public      => true,
+          :type        => Array,
+          :allowed_from_server => false,
+          :transform    => DefaultSource.method(:convert_to_list),
+          :description => 'Prefix of attributes to include on span events. Allows <code>*</code> as wildcard at end.'
+        },
+        :'transaction_segments.attributes.include' => {
+          :default     => [],
+          :public      => true,
+          :type        => Array,
+          :allowed_from_server => false,
+          :transform    => DefaultSource.method(:convert_to_list),
+          :description => 'Prefix of attributes to include on transaction segments. Allows <code>*</code> as wildcard at end.'
         },
         :'custom_attributes.enabled' => {
           :default     => true,