+++ lib/new_relic/agent/transaction/trace_node.rb
@@ -21,8 +21,10 @@ module NewRelic
           @entry_timestamp = relative_start
           @metric_name     = metric_name || UNKNOWN_NODE_NAME
           @exit_timestamp  = relative_end
-          @children    = nil
-          @params          = params
+          @children        = nil
+          @params          = params.select do |p|
+            NewRelic::Agent.instance.attribute_filter.allows_key? p, AttributeFilter::DST_TRANSACTION_SEGMENTS
+          end if params
           @parent_node     = parent
         end